requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.12.5
chardet==4.0.0
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-extra-fields==3.0.2
djangorestframework==3.12.2
idna==2.10
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
psycopg2 @ file:///C:/ci/psycopg2_1612298803839/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests==2.25.1
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
torch==1.7.1
torchvision==0.2.2.post3
typing-extensions==3.7.4.3
urllib3==1.26.3
wincertstore==0.2
